Fix 32bit wrap to att buffer size (#176)

Everything but PSDB passed. I'm force merging because 1) this is a minor change and 2) PSDBs seem broken now, they are stuck and pointing to the wrong location: http://rocm-ci.amd.com/job/compute-psdb-staging-profiler-emu/435/ 

[ROCm/aqlprofile commit: cec481b4b1a223c4dd8430d079bd27ca66da3303]

+++ projects/aqlprofile/gfxip/gfx9/gfx9_primitives.h
@@ -683,7 +683,7 @@ class gfx9_cntx_prim {
 
   // Indicates the size of buffer to use per Shader Engine instance.
   // The size is specified in terms of 4KB blocks
-  static uint32_t sqtt_buffer_size_value(uint32_t size_val, uint32_t base_hi) {
+  static uint32_t sqtt_buffer_size_value(uint64_t size_val, uint32_t base_hi) {
     uint32_t sq_thread_trace_size =
         SET_REG_FIELD_BITS(SQ_THREAD_TRACE_SIZE, SIZE, (size_val >> TT_BUFF_ALIGN_SHIFT));
     return sq_thread_trace_size;
